fix entity services and controllers, add all relation services and controllers, add Postman tests

+++ src/artwork/artwork.service.ts
@@ -21,37 +21,48 @@ export class ArtworkService {
     private readonly artistRepository: Repository<Artist>,
   ) {}
 
-  async findAll(): Promise<ArtworkDTO[]> {
-    return await this.artworkRepository.find({ relations : ['exhibition', 'museum', 'images', 'artist'] });
+  async findAll(artistId: number): Promise<ArtworkDTO[]> {
+    const artist = await this.artistRepository.findOne(artistId, { relations : ['artworks'] });
+    if (!artist)
+      throw new BusinessLogicException("The artist with the given id was not found", BusinessError.NOT_FOUND);
+    return artist.artworks;
   }
 
-  async findOne(id: number): Promise<ArtworkDTO> {
-    const artwork = await this.artworkRepository.findOne(id, { relations: ['exhibition', 'museum', 'images', 'artist'] });
+  async findOne(artistId: number, artworkId: number): Promise<ArtworkDTO> {
+    const artist = await this.artistRepository.findOne(artistId, { relations: ["artworks"] });
+    if (!artist)
+      throw new BusinessLogicException("The artist with the given id was not found", BusinessError.NOT_FOUND);
+
+    const artwork = await this.artworkRepository.findOne(artworkId, { relations: ["artist"] });
     if (!artwork)
       throw new BusinessLogicException("The artwork with the given id was not found", BusinessError.NOT_FOUND)
-    else
-      return artwork;
+    
+    const artworkartist = artist.artworks.find(e => e.id === artwork.id);
+    if (!artworkartist)
+      throw new BusinessLogicException("The artwork is not associated to the artist", BusinessError.NOT_FOUND)
+
+    return artworkartist;
   }
 
-  async create(artworkDTO: ArtworkDTO): Promise<ArtworkDTO> {
-    if (artworkDTO.museum == null)
-      throw new BusinessLogicException("The artwork must have a museum association", BusinessError.PRECONDITION_FAILED);
+  async create(artistId: number, artworkDTO: ArtworkDTO): Promise<ArtworkDTO> {
+    var museum = null;
+    if (artworkDTO.museum != null) {
+      museum = await this.museumRepository.findOne(artworkDTO.museum.id);
+      if (!museum)
+        throw new BusinessLogicException("The museum with the given id was not found", BusinessError.NOT_FOUND);
+    }
 
-    const museum = await this.museumRepository.findOne(artworkDTO.museum.id);
-    if (!museum)
-      throw new BusinessLogicException("The museum with the given id was not found", BusinessError.NOT_FOUND);
+    var exhibition = null;
+    if (artworkDTO.exhibition != null) {
+      exhibition = await this.exhibitionRepository.findOne(artworkDTO.exhibition.id);
+      if (!exhibition)
+        throw new BusinessLogicException("The exhibition with the given id was not found", BusinessError.NOT_FOUND);
+    }
 
-    if (artworkDTO.exhibition == null)
-      throw new BusinessLogicException("The artwork must have an exhibition association", BusinessError.PRECONDITION_FAILED);
-
-    const exhibition = await this.exhibitionRepository.findOne(artworkDTO.exhibition.id);
-    if (!exhibition)
-      throw new BusinessLogicException("The exhibition with the given id was not found", BusinessError.NOT_FOUND);
-    
     if (artworkDTO.artist == null)
       throw new BusinessLogicException("The artwork must have an artist association", BusinessError.PRECONDITION_FAILED);
 
-    const artist = await this.artistRepository.findOne(artworkDTO.artist.id);
+    const artist = await this.artistRepository.findOne(artistId);
     if (!artist)
       throw new BusinessLogicException("The artist with the given id was not found", BusinessError.NOT_FOUND);
 
@@ -66,29 +77,29 @@ export class ArtworkService {
     return await this.artworkRepository.save(artwork);
   }
 
-  async update(id: number, artworkDTO: ArtworkDTO): Promise<ArtworkDTO> {
-    const artwork = await this.artworkRepository.findOne(id);
+  async update(artistId: number, artworkId: number, artworkDTO: ArtworkDTO): Promise<ArtworkDTO> {
+    const artwork = await this.artworkRepository.findOne(artworkId);
     if (!artwork)
       throw new BusinessLogicException("The artwork with the given id was not found", BusinessError.NOT_FOUND)
     
-    if (artworkDTO.museum == null)
-      throw new BusinessLogicException("The artwork must have a museum association", BusinessError.PRECONDITION_FAILED);
-
-    const museum = await this.museumRepository.findOne(artworkDTO.museum.id);
-    if (!museum)
-      throw new BusinessLogicException("The museum with the given id was not found", BusinessError.NOT_FOUND);
-
-    if (artworkDTO.exhibition == null)
-      throw new BusinessLogicException("The artwork must have an exhibition association", BusinessError.PRECONDITION_FAILED);
-
-    const exhibition = await this.exhibitionRepository.findOne(artworkDTO.exhibition.id);
-    if (!exhibition)
-      throw new BusinessLogicException("The exhibition with the given id was not found", BusinessError.NOT_FOUND);
+    var museum = null;
+    if (artworkDTO.museum != null) {
+      museum = await this.museumRepository.findOne(artworkDTO.museum.id);
+      if (!museum)
+        throw new BusinessLogicException("The museum with the given id was not found", BusinessError.NOT_FOUND);
+    }
+  
+    var exhibition = null;
+    if (artworkDTO.exhibition != null) {
+      exhibition = await this.exhibitionRepository.findOne(artworkDTO.exhibition.id);
+      if (!exhibition)
+        throw new BusinessLogicException("The exhibition with the given id was not found", BusinessError.NOT_FOUND);
+    }
 
     if (artworkDTO.artist == null)
       throw new BusinessLogicException("The artwork must have an artist association", BusinessError.PRECONDITION_FAILED);
 
-    const artist = await this.artistRepository.findOne(artworkDTO.artist.id);
+    const artist = await this.artistRepository.findOne(artistId);
     if (!artist)
       throw new BusinessLogicException("The artist with the given id was not found", BusinessError.NOT_FOUND);
 
@@ -104,11 +115,17 @@ export class ArtworkService {
     return artwork;
   }
 
-  async delete(id: number) {
-    const artwork = await this.artworkRepository.findOne(id);
+  async delete(artistId: number, artworkId: number) {
+    const artist = await this.artistRepository.findOne(artistId, { relations: ['artworks'] });
+    if (!artist)
+      throw new BusinessLogicException("The artist with the given id was not found", BusinessError.NOT_FOUND)
+
+    const artwork = await this.artworkRepository.findOne(artworkId);
     if (!artwork)
       throw new BusinessLogicException("The artwork with the given id was not found", BusinessError.NOT_FOUND)
-    else
-      return await this.artworkRepository.remove(artwork);
+    
+    artist.artworks = artist.artworks.filter(e => e.id !== artwork.id);
+    await this.artistRepository.save(artist);
+    return await this.artworkRepository.remove(artwork);
   }
 }
